Implementado manejo de excepciones para validar la entrada del usuario: número positivo y formato correcto.tambien modifique y agregue  una lineas de codigo q me hacian me hacian falta en calculo_numero.py
