+++ src/calculo_numeros.py
@@ -1,5 +1,24 @@
-from exceptions import ingrese_numero
+from exceptions import NumeroDebeSerPositivo
 
+def ingrese_numero():
+    """
+    Solicita al usuario ingresar un número y valida que sea positivo.
+    
+    Returns:
+        int: El número ingresado si es válido.
+        
+    Raises:
+        ValueError: Si la entrada no es un número válido.
+        NumeroDebeSerPositivo: Si el número ingresado es negativo.
+    """
+    entrada = input("Ingrese un número: ")
+    try:
+        numero = int(entrada)
+        if numero < 0:
+            raise NumeroDebeSerPositivo()
+        return numero
+    except ValueError:
+        raise ValueError("La entrada debe ser un número válido")
 def main():
     """
     Programa principal que solicita números al usuario y muestra los resultados.
